Move StorageLink to common
